+++ examples/todomvc/Redux.TodoMvc.Android/MainActivity.cs
@@ -9,8 +9,7 @@ namespace Redux.TodoMvc.Android
     [Activity(Label = "Redux.TodoMvc.Android", MainLauncher = true, Icon = "@drawable/icon")]
     public class MainActivity : Activity
     {
-        public  IStore<ApplicationState> Store { get; private set; }
-
+        public IStore<ApplicationState> Store { get; private set; }
 
         public MainActivity()
         {
@@ -28,7 +27,6 @@ namespace Redux.TodoMvc.Android
             base.OnCreate(bundle);
             SetContentView(Resource.Layout.Main);
         }
-        
     }
 }
