+++ examples/todomvc/Redux.TodoMvc.Android/MainSection.cs
@@ -5,7 +5,6 @@ using Android.App;
 using Android.OS;
 using Android.Views;
 using Android.Widget;
-using Redux.TodoMvc.Actions;
 using Redux.TodoMvc.States;
 
 namespace Redux.TodoMvc.Android
@@ -14,6 +13,8 @@ namespace Redux.TodoMvc.Android
     {
         private ListView _listView;
         private View _view;
+
+        public IStore<ApplicationState> ActivityStore { get; set; }
 
         public override void OnCreate(Bundle savedInstanceState)
         {
@@ -27,6 +28,13 @@ namespace Redux.TodoMvc.Android
 
                 _listView.Adapter = new ListItemAdapter(this.Activity, list, ActivityStore);
             });
+        }
+
+        public override View OnCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState)
+        {
+            _view = inflater.Inflate(Resource.Layout.MainSection, container, false);
+            _listView = _view.FindViewById<ListView>(Resource.Id.listView1);
+            return _view;
         }
 
         private List<Todo> FilterTodos(ApplicationState applicationState)
@@ -47,65 +55,6 @@ namespace Redux.TodoMvc.Android
             }
 
             return todos;
-        }
-
-        public IStore<ApplicationState> ActivityStore { get; set; }
-
-        public override View OnCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState)
-        {
-            _view = inflater.Inflate(Resource.Layout.MainSection, container, false);
-            _listView = _view.FindViewById<ListView>(Resource.Id.listView1);
-            return _view;
-        }
-    }
-
-    public class ListItemAdapter : BaseAdapter<Todo>
-    {
-        private readonly List<Todo> _items;
-        private readonly IStore<ApplicationState> _activityStore;
-        private readonly Activity _context;
-        public ListItemAdapter(Activity context, List<Todo> items, IStore<ApplicationState> activityStore)
-        {
-            _context = context;
-            _items = items;
-            _activityStore = activityStore;
-        }
-        public override long GetItemId(int position)
-        {
-            return position;
-        }
-        public override Todo this[int position] => _items[position];
-
-        public override int Count => _items.Count;
-
-        public override View GetView(int position, View convertView, ViewGroup parent)
-        {
-            var item = _items[position];
-            var view = _context.LayoutInflater.Inflate(Resource.Layout.ListViewItem, null);
-
-            var checkBox = view.FindViewById<CheckBox>(Resource.Id.checkbox);
-
-            checkBox.Checked = item.IsCompleted;
-            checkBox.CheckedChange += delegate
-            {
-                _activityStore.Dispatch(new CompleteTodoAction()
-                {
-                    TodoId = item.Id
-                });
-            };
-
-
-            view.FindViewById<TextView>(Resource.Id.todoTextItem).Text = item.Text;
-
-            view.FindViewById<Button>(Resource.Id.RemoveButton).Click += delegate
-            {
-                _activityStore.Dispatch(new DeleteTodoAction
-                {
-                    TodoId = item.Id
-                });
-            };
-
-            return view;
         }
     }
 }
