+++ examples/todomvc/Redux.TodoMvc.Android/Resources/layout/MainSection.xml
@@ -12,4 +12,5 @@
        android:layout_width="match_parent"
        android:layout_height="match_parent"
        android:id="@+id/listView1" />
+  
 </LinearLayout>
